Refactor: keyboard Event hook으로 분리
What is this PR?
- useHandleKeyBoard 커스텀 훅으로 분리
