+++ client/src/components/card/feedwritecard/FeedWriteCard.tsx
@@ -1,4 +1,4 @@
-import { useState, useCallback, useRef, useEffect } from 'react';
+import { useState, useCallback, useRef } from 'react';
 import { useParams, useNavigate } from 'react-router-dom';
 import { Navigation } from 'swiper/modules';
 import { Swiper, SwiperSlide } from 'swiper/react';
@@ -22,6 +22,7 @@ import NoticeServerError from '@/components/notice/NoticeServerError.tsx';
 import usePatchFormMutation from '@/hook/api/mutation/usePatchFormMutation';
 import usePostFormMutation from '@/hook/api/mutation/usePostFormMutation';
 import useFeedDetailQuery from '@/hook/api/query/useFeedDetailQuery';
+import useHandleKeyBoard from '@/hook/useHandleKeyBoard';
 import Path from '@/routers/paths.ts';
 import { FeedImage } from '@/types/feedTypes.ts';
 import { parseImg, deleteImg } from '@/util/parseImg.ts';
@@ -50,18 +51,9 @@ export default function FeedWriteCard() {
   const accessToken = useReadLocalStorage<string>('accessToken');
 
   // esc 누를 시, 창닫기
-  useEffect(() => {
-    const getBackPage = (e: KeyboardEvent) => {
-      if (e.key === 'Escape') {
-        navigate(-1);
-      }
-    };
-    window.addEventListener('keydown', getBackPage);
-    return () => window.removeEventListener('keydown', getBackPage);
-  }, [navigate]);
+  useHandleKeyBoard({ navigate });
 
   // param이 있을 경우,  피드 게시물 정보 가져오기
-
   const getFeedDetailQuery = useFeedDetailQuery({
     feedId: Number(feedId),
     url: `${SERVER_URL}/feeds/${feedId}`,
@@ -141,93 +133,95 @@ export default function FeedWriteCard() {
         <NoticeServerError />
       </div>
     );
+
   if (
     getFeedDetailQuery.isFetching &&
     feedPostingMutation.isLoading &&
     feedEditingMutation.isLoading
   ) {
     return <LoadingComponent />;
-  } else
-    return (
-      <>
-        {isOpen[0] && (
-          <Popup
-            title={isOpen[1]}
-            handler={[() => setIsOpen([false, ''])]}
-            btnsize={['md']}
-            countbtn={1}
-            buttontext={['확인']}
-            isgreen={['true']}
-            popupcontrol={() => setIsOpen([false, ''])}
+  }
+
+  return (
+    <>
+      {isOpen[0] && (
+        <Popup
+          title={isOpen[1]}
+          handler={[() => setIsOpen([false, ''])]}
+          btnsize={['md']}
+          countbtn={1}
+          buttontext={['확인']}
+          isgreen={['true']}
+          popupcontrol={() => setIsOpen([false, ''])}
+        />
+      )}
+      <div
+        className="w-screen h-screen bg-zinc-900/75  absolute flex items-center justify-center max-sm:items-start max-sm:pt-10"
+        onClick={e => {
+          if (e.target === e.currentTarget) navigate(-1);
+        }}>
+        <Container>
+          <Close
+            className="absolute right-6 top-6 cursor-pointer"
+            fill="black"
+            onClick={() => navigate(-1)}
           />
-        )}
-        <div
-          className="w-screen h-screen bg-zinc-900/75  absolute flex items-center justify-center max-sm:items-start max-sm:pt-10"
-          onClick={e => {
-            if (e.target === e.currentTarget) navigate(-1);
-          }}>
-          <Container>
-            <Close
-              className="absolute right-6 top-6 cursor-pointer"
-              fill="black"
-              onClick={() => navigate(-1)}
-            />
-            <Wrap>
-              <Title>게시물 작성</Title>
-            </Wrap>
+          <Wrap>
+            <Title>게시물 작성</Title>
+          </Wrap>
 
-            <Swiper
-              pagination={{
-                type: 'fraction',
-              }}
-              navigation={true}
-              modules={[Navigation]}
-              className="w-80 h-[360px] rounded-lg">
-              {savedFile.map((file, idx) => (
-                <SwiperSlide
-                  key={idx}
-                  className="bg-lightgray flex justify-center items-center relative rounded-lg">
-                  <img
-                    src={file}
-                    className="rounded-lg border border-lightgray"
+          <Swiper
+            pagination={{
+              type: 'fraction',
+            }}
+            navigation={true}
+            modules={[Navigation]}
+            className="w-80 h-[360px] rounded-lg">
+            {savedFile.map((file, idx) => (
+              <SwiperSlide
+                key={idx}
+                className="bg-lightgray flex justify-center items-center relative rounded-lg">
+                <img
+                  src={file}
+                  className="rounded-lg border border-lightgray"
+                />
+                <Close
+                  className="absolute top-3 right-3 cursor-pointer"
+                  fill="#a1a1aa"
+                  onClick={() => handleSemiClose(idx)}
+                />
+              </SwiperSlide>
+            ))}
+
+            {savedFile.length < 3 && (
+              <SwiperSlide>
+                <Form>
+                  <Pluslabel htmlFor="feedimg">
+                    <Plus className="w-full mt-[180px]" />
+                  </Pluslabel>
+                  <input
+                    id="feedimg"
+                    type="file"
+                    accept="image/*"
+                    onChange={e => handleUpload(e)}
+                    style={{ display: 'none' }}
+                    multiple
                   />
-                  <Close
-                    className="absolute top-3 right-3 cursor-pointer"
-                    fill="#a1a1aa"
-                    onClick={() => handleSemiClose(idx)}
-                  />
-                </SwiperSlide>
-              ))}
+                </Form>
+              </SwiperSlide>
+            )}
+          </Swiper>
 
-              {savedFile.length < 3 && (
-                <SwiperSlide>
-                  <Form>
-                    <Pluslabel htmlFor="feedimg">
-                      <Plus className="w-full mt-[180px]" />
-                    </Pluslabel>
-                    <input
-                      id="feedimg"
-                      type="file"
-                      accept="image/*"
-                      onChange={e => handleUpload(e)}
-                      style={{ display: 'none' }}
-                      multiple
-                    />
-                  </Form>
-                </SwiperSlide>
-              )}
-            </Swiper>
-
-            <Textarea
-              placeholder="글을 입력해주세요."
-              maxLength={200}
-              ref={textRef}
-            />
-            <SubmitBtn type="button" onClick={handleSubmit}>
-              <Write />
-            </SubmitBtn>
-          </Container>
-        </div>
-      </>
-    );
+          <Textarea
+            placeholder="글을 입력해주세요."
+            maxLength={200}
+            ref={textRef}
+          />
+          <SubmitBtn type="button" onClick={handleSubmit}>
+            <Write />
+          </SubmitBtn>
+        </Container>
+      </div>
+    </>
+  );
 }
