+++ client/src/util/upload.ts
@@ -1,0 +1,4 @@
+//  FeedCardUpload에 사용
+export const handleFeedCardUpload = () => {
+  //
+};
